pipeline: Move source map related login into `sources.rs`
